move functions to utils and better command options
